refactor: serverId 로컬 상태 삭제, Template 컴포넌트 위치 변경
