+++ src/components/index.js
@@ -27,6 +27,3 @@ export { Select } from './atoms/Select';
 export { NavBar } from './organisms/NavBar';
 export { ServerSelect } from './organisms/ServerSelect';
 export { ServerSelectItem } from './organisms/ServerSelectItem';
-export { StationListItem } from './organisms/StationListItem';
-
-export { Page } from './templates/Page';
